# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Model Computer Vision API",
...
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/classify": {
      "post": {
        "tags": [
          "Classification"
        ],
        "summary": "Classify an image (JSON body)",
        "description": "Classify a single image supplied as base64 or a URL.\n\nSend **either** ``image_base64`` **or** ``image_url``, never both.\n\nExample request::\n\n    {\n      \"image_base64\": \"iVBORw0KGgoAAA...\",\n      \"top_k\": 5,\n      \"confidence_threshold\": 0.1\n    }\n\nExample response::\n\n    {\n      \"predictions\": [\n        {\"class_id\": 207, \"label\": \"golden retriever\", \"confidence\": 0.93, \"rank\": 1}\n      ],\n      \"model\": {\"name\": \"resnet50\", \"version\": \"1.0.0\", \"runtime\": \"onnx\", \"device\": \"cpu\"},\n      \"timing\": {\"preprocess_ms\": 8.1, \"inference_ms\": 31.4, \"total_ms\": 41.9},\n      \"correlation_id\": \"0f6c...\",\n      \"cached\": false\n    }",
        "operationId": "classify_api_v1_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "The image could not be decoded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The image exceeds the size limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for your tier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The model is not ready to serve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Inference exceeded its deadline.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classify/upload": {
      "post": {
        "tags": [
          "Classification"
        ],
        "summary": "Classify an uploaded image file (multipart)",
        "description": "Classify an image sent as a multipart file upload.\n\nThis is the endpoint to use from a browser form or with::\n\n    curl -X POST http://localhost:8000/api/v1/classify/upload \\\n         -H \"X-API-Key: your-key\" \\\n         -F \"file=@cat.jpg\" -F \"top_k=3\"",
        "operationId": "classify_upload_api_v1_classify_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_classify_upload_api_v1_classify_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "The image could not be decoded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The image exceeds the size limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for your tier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The model is not ready to serve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Inference exceeded its deadline.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_classify_upload_api_v1_classify_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Image file to classify."
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "confidence_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold",
            "default": 0.0
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeFormat"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_classify_upload_api_v1_classify_upload_post"
      },
...
      "ClassPrediction": {
        "properties": {
          "class_id": {
            "type": "integer",
            "title": "Class Id",
            "description": "Zero-based index of the class in the model's label set."
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Human-readable class name."
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Softmax probability, 0 to 1."
          },
          "rank": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rank",
            "description": "1 = highest scoring."
          }
        },
        "type": "object",
        "required": [
          "class_id",
          "label",
          "confidence",
          "rank"
        ],
        "title": "ClassPrediction",
        "description": "A single predicted class."
      },
      "ClassificationRequest": {
        "properties": {
          "model_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-zA-Z0-9_\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "Registry model name. Defaults to the task's default model."
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "pattern": "^[a-zA-Z0-9_.\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version",
            "description": "Registry version, or 'latest'. Defaults to 'latest'."
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeFormat"
              },
              {
                "type": "null"
              }
            ],
            "description": "Force a specific runtime format. Defaults to the server's preferred runtime."
          },
          "image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Base64",
            "description": "Raw image bytes, base64-encoded. A `data:` URI prefix is accepted and stripped."
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Publicly reachable HTTPS URL to fetch the image from."
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id",
            "description": "Caller-supplied identifier echoed back in the response, useful for batches."
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "How many of the highest-scoring classes to return.",
            "default": 5
          },
          "confidence_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Threshold",
            "description": "Drop predictions scoring below this probability.",
            "default": 0.0
          },
          "include_probabilities": {
            "type": "boolean",
            "title": "Include Probabilities",
            "description": "Include the full softmax probability for each class.",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ClassificationRequest",
        "description": "Body for ``POST /api/v1/classify``.",
        "examples": [
          {
            "confidence_threshold": 0.1,
            "image_base64": "iVBORw0KGgoAAAANSUhEUgAA...",
            "top_k": 5
          }
        ]
      },
      "ClassificationResponse": {
        "properties": {
          "model": {
            "$ref": "#/components/schemas/ModelInfo"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingInfo"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id",
            "description": "Quote this id in any support request."
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "True when this result was served from Redis instead of re-running the model.",
            "default": false
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id",
            "description": "Echo of the caller-supplied image_id, if any."
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "description": "True when the primary model was unavailable and a fallback model or runtime served the request instead.",
            "default": false
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Non-fatal notes, e.g. 'image was resized' or 'fell back to torch runtime'."
          },
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/ClassPrediction"
            },
            "type": "array",
            "title": "Predictions",
            "description": "Top-k predictions, highest confidence first."
          },
          "top_prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassPrediction"
              },
              {
                "type": "null"
              }
            ],
            "description": "Convenience copy of predictions[0]; null if the threshold filtered everything out."
          }
        },
        "type": "object",
        "required": [
          "model",
          "timing",
          "correlation_id",
          "predictions"
        ],
        "title": "ClassificationResponse",
        "description": "Response for ``POST /api/v1/classify``."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable code, e.g. 'IMAGE_TOO_LARGE'."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Friendly, user-facing explanation."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Inner object of the standard error envelope."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The single error shape every failing endpoint returns."
      },
...
      "ModelInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Registry model name, e.g. 'resnet50-tiny-imagenet'."
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Registry version, e.g. '1.0.0'."
          },
          "task": {
            "$ref": "#/components/schemas/TaskType",
            "description": "The task this model performs."
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeFormat",
            "description": "Which compiled format actually executed."
          },
          "device": {
            "type": "string",
            "title": "Device",
            "description": "'cpu' or 'cuda'."
          }
        },
        "type": "object",
        "required": [
          "name",
          "version",
          "task",
          "runtime",
          "device"
        ],
        "title": "ModelInfo",
        "description": "Which model produced a result. Present on every inference response."
      },
...
      "RuntimeFormat": {
        "type": "string",
        "enum": [
          "torch",
          "torch_int8",
          "onnx",
          "onnx_int8",
          "tensorrt"
        ],
        "title": "RuntimeFormat",
        "description": "Which compiled form of a model is executing."
      },
...
      "TaskType": {
        "type": "string",
        "enum": [
          "classification",
          "detection",
          "similarity"
        ],
        "title": "TaskType",
        "description": "The three computer-vision tasks this system serves."
      },
      "TimingInfo": {
        "properties": {
          "preprocess_ms": {
            "type": "number",
            "title": "Preprocess Ms",
            "description": "Decode, validate, resize and normalise."
          },
          "inference_ms": {
            "type": "number",
            "title": "Inference Ms",
            "description": "Time inside the model's forward pass."
          },
          "postprocess_ms": {
            "type": "number",
            "title": "Postprocess Ms",
            "description": "Turning raw tensors into JSON-ready results."
          },
          "total_ms": {
            "type": "number",
            "title": "Total Ms",
            "description": "End-to-end time spent inside the handler."
          }
        },
        "type": "object",
        "required": [
          "preprocess_ms",
          "inference_ms",
          "postprocess_ms",
          "total_ms"
        ],
        "title": "TimingInfo",
        "description": "Latency breakdown in milliseconds, for performance debugging.\n\nSplitting preprocessing from inference matters: if total latency rises,\nthese numbers say immediately whether the model got slower or whether the\nsystem is spending its time decoding and resizing images."
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key issued to your account. Determines your tier."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT access token."
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
}